pyproject.toml
[tool.poetry]
name = "subject-screener"
version = "1.0.0"
description = "App for project"
authors = ["SRL"]
readme = "README.md"
packages = [{include = "sub_search"}, {"include"= "sub_process"}]


[tool.poetry.dependencies]
python = "3.8.5"
transformers = "^4.34.0"
apache-airflow = "^2.7.2"
sub-search = {path = "sub_search"}
sub-process = {path = "sub_process"}
google-cloud-bigquery = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
